+++ R/grouped_gghistostats.R
@@ -12,31 +12,39 @@
 #'   taken.
 #' @param x A numeric variable.
 #' @param xlab Label for `x` axis variable.
+#' @param title The text for the plot title.
 #' @param subtitle The text for the plot subtitle *if* you don't want results
 #'   from one sample test to be displayed.
 #' @param caption The text for the plot caption.
-#' @param type Type of statistic expected (`"parametric"` or `"nonparametric"` or
-#'   `"bayes"`). Abbreviations accepted are `"p"` or `"np"` or `"bf"`, respectively.
+#' @param type Type of statistic expected (`"parametric"` or `"nonparametric"`
+#'   or `"bayes"`). Abbreviations accepted are `"p"` or `"np"` or `"bf"`,
+#'   respectively.
 #' @param test.value A number specifying the value of the null hypothesis.
-#' @param bf.prior A number between 0.5 and 2 (default `0.707`), the prior width
+#' @param bf.prior A number between 0.5 and 2 (default 0.707), the prior width
 #'   to use in calculating Bayes factors.
 #' @param bf.message Logical. Decides whether to display Bayes Factor in favor
 #'   of null hypothesis for parametric test if the null hypothesis can't be
 #'   rejected (Default: `bf.message = TRUE`).
 #' @param k Number of decimal places expected for results.
+#' @param low.color,high.color Colors for low and high ends of the gradient.
+#'   Defaults are colorblind-friendly.
 #' @param results.subtitle Decides whether the results of statistical tests are
 #'   to be displayed as subtitle (Default: `results.subtitle = TRUE`). If set to
 #'   `FALSE`, no statistical tests will be run.
-#' @param centrality.para Decides *which* measure of central tendency (`"mean"` or
-#'   `"median"`) is to be displayed as a vertical line.
-#' @param centrality.colour Decides colour for the vertical line for centrality
+#' @param legend.title.margin Adjusting the margin between legend title and the
+#'   colorbar.
+#' @param t.margin,b.margin Margins in grid units. For more details, see
+#'   `?grid::unit()`.
+#' @param centrality.para Decides *which* measure of central tendency (`"mean"`
+#'   or `"median"`) is to be displayed as a vertical line.
+#' @param centrality.color Decides color for the vertical line for centrality
 #'   parameter (Default: `"blue"`).
 #' @param test.value.line Decides whether test value is to be displayed as a
 #'   vertical line (Default: `FALSE`).
-#' @param test.value.colour Decides colour for the vertical line denoting test
+#' @param test.value.color Decides color for the vertical line denoting test
 #'   value (Default: `"black"`).
-#' @param binwidth.adjust If set to `TRUE`, you can use it to pick better value
-#'   with the `binwidth` argument to `stat_bin()`.
+#' @param line.labeller A logical that decides whether line labels should be
+#'   displayed (Default: `FALSE`).
 #' @param binwidth The width of the bins. Can be specified as a numeric value,
 #'   or a function that calculates width from `x`. The default is to use bins
 #'   bins that cover the range of the data. You should always override this
@@ -73,47 +81,6 @@
 #' @export
 #'
 
-# defining global variables and functions to quient the R CMD check notes
-utils::globalVariables(
-  c(
-    "%<>%",
-    "U",
-    "V",
-    "Z",
-    "chi",
-    "counts",
-    "df",
-    "df1",
-    "df2",
-    "effsize",
-    "estimate",
-    "eta",
-    "omega",
-    "perc",
-    "cramer",
-    "pvalue",
-    "r",
-    "rho",
-    "xi",
-    "y",
-    "z_value",
-    "italic",
-    "rsubtitle",
-    "stats_subtitle",
-    "chi_subtitle",
-    "proptest_subtitle",
-    "LL",
-    "UL",
-    "..count..",
-    "dnorm",
-    "mean",
-    "median",
-    "sd",
-    "bf",
-    "bf_error"
-  )
-)
-
 # defining the function
 grouped_gghistostats <- function(grouping.var,
                                  data,
@@ -127,11 +94,14 @@ grouped_gghistostats <- function(grouping.var,
                                  bf.message = TRUE,
                                  k = 3,
                                  results.subtitle = TRUE,
+                                 legend.title.margin = TRUE,
+                                 t.margin = unit(0, "mm"),
+                                 b.margin = unit(3, "mm"),
                                  centrality.para = NULL,
-                                 centrality.colour = "blue",
+                                 centrality.color = "blue",
                                  test.value.line = FALSE,
-                                 test.value.colour = "black",
-                                 binwidth.adjust = FALSE,
+                                 test.value.color = "black",
+                                 line.labeller = FALSE,
                                  binwidth = NULL,
                                  messages = TRUE,
                                  ...) {
@@ -171,10 +141,9 @@ grouped_gghistostats <- function(grouping.var,
             k = k,
             results.subtitle = results.subtitle,
             centrality.para = centrality.para,
-            centrality.colour = centrality.colour,
+            centrality.color = centrality.color,
             test.value.line = test.value.line,
-            test.value.colour = test.value.colour,
-            binwidth.adjust = binwidth.adjust,
+            test.value.color = test.value.color,
             binwidth = binwidth,
             messages = messages
           )