+++ R/theme_mprl.R
@@ -34,7 +34,7 @@ theme_mprl <- function() {
       legend.key.width = grid::unit(x = 1, units = "line"),
       plot.margin = grid::unit(x = c(1, 1, 1, 1), units = "lines"),
       panel.border = ggplot2::element_rect(
-        colour = "black",
+        color = "black",
         fill = NA,
         size = 1
       ),